requirements.txt
asgiref==3.4.1
dict2xml==1.7.0
dicttoxml==1.7.4
Django==3.2.9
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
mysqlclient==2.0.3
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
